# Faster move generation and execution