This PR:

1. Relocates the `move_actions` and `revise_move` logic from `LocalGame` to `reconchess.utilities` for easier use by bot developers and
2. Optimizes those two functions and `slide_move`, which gives a ~5x speedup in random moving (example script below).

To test the performance, I ran the following

```
from reconchess.utilities import *
from chess import *

import random

def run(n=1_000):
    board = Board()
    for _ in range(n):
        move = random.choice(move_actions(board))
        revised = revise_move(board, move) or Move.null()
        board.push(revised)
        if board.king(WHITE) is None or board.king(BLACK) is None:
            board = Board()
    return board
```

On master you have to define the following (copied from LocalGame with no changes)

```
def move_actions(board):
    return moves_without_opponent_pieces(board) + pawn_capture_moves_on(board)

def revise_move(board, move):
    # if its a legal move, don't change it at all. note that board.generate_psuedo_legal_moves() does not
    # include psuedo legal castles
    if move in board.generate_pseudo_legal_moves() or is_psuedo_legal_castle(board, move):
        return move

    # note: if there are pieces in the way, we DONT capture them
    if is_illegal_castle(board, move):
        return None

    # if the piece is a sliding piece, slide it as far as it can go
    piece = board.piece_at(move.from_square)
    if piece.piece_type in [chess.PAWN, chess.ROOK, chess.BISHOP, chess.QUEEN]:
        move = slide_move(board, move)

    return move if move in board.generate_pseudo_legal_moves() else None
```

Before changes

```
%time random.seed(12890); run(100_000)
CPU times: user 53.9 s, sys: 190 ms, total: 54 s
Wall time: 55.8 s
```

After changes
```
%time random.seed(12890); run(100_000)
CPU times: user 10.6 s, sys: 79.7 ms, total: 10.7 s
Wall time: 11 s
```